remove antiquated add places to firestore and improve logging for various errors on startup
